requirements.txt
faiss
sklearn
indic_nlp_library
indic_nlp_resources